ADD: UPD: extend impl classes of QueryForm to extend from AbstractQueryForm
