+++ sparql/src/main/java/de/uni/stuttgart/ipvs/sparql/query/AskQuery.java
@@ -1,0 +1,9 @@
+package de.uni.stuttgart.ipvs.sparql.query;
+
+public class AskQuery extends AbstractQueryForm {
+
+    @Override
+    public String getString() {
+        return "ASK " + super.getString();
+    }
+}
